requirements.txt
bootstrap4
django
django-crispy-forms
djangorestframework
pandas